menu_reformer/requirements.txt
attrs==22.1.0
boto3==1.26.33
botocore==1.29.33
certifi==2022.12.7
charset-normalizer==2.1.1
exceptiongroup==1.0.4
idna==3.4
iniconfig==1.1.1
jmespath==1.0.1
packaging==22.0
pluggy==1.0.0
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
pyjwt==2.6.0
jsonpath-ng==1.5.3